Organized files and implemented an early version of a.i.m.
Artificial Intelligence Manager has the ability to execute commands to
help
manage your personal AI. So far there are only the commands help and
memories. help will return all aim commands and memories will return a
memory tree for the entered memory (/all to return a master tree).
Future vision is to enable an admin for the AI with special commands and
a sort of
"download memories" system. Such as an "AI" package (manager) !